Author SQL file for adding dummy data to database
One can now use dummy.sql to populate the database
with some dummy data.  The schema creation file is also
updated with necessary permissions to insert without
getting slapped by the auto-increment id field.

The dummy data file is expected to grow and be updated
as time goes on and fresher data is required for testing.