Update calendar view to display dummy data
The calendar view has been adapted from an older app to
visualize flavourless json fetched from the database.  The
database query for this view only asks for the days which
have datapoints associated with them for the current month,
with no regard for how many datapoints each day has, or
what kind of data is associated with said day.

This month view is wrapped in a div which can be made
smaller, so pending some changes should serve well as a
basis for a year view, which will be accessed using the
calendar view, only giving one parameter (the year)
instead of two (year, month).