Add view for listing entries by date
A view for listing entries by date, generally accessed
via the calendar view, is now available.  Some style
tweaks come with making this view look OK.

High importance: data validation is required, but may be
deferred until the first tests are written so that this
can be done in a uniform way.