Implement edit and new entry views for flat data
The views for viewing and editing data entries now display,
with some limitations which will need to be addressed soon.
Most importantly, these views only work for JSON that is
single-level.  Also, there is no support for setting
seconds, milliseconds, or timezone information yet.  Saving
is not done yet, as that will need to be its own commit.