Implement edit view for multi-level JSON
The edit view can now handle displaying multi-level JSON
thanks to box-drawing characters to indicate hierarchy.  It
still doesn't save data, and the means for adding more
fields is now broken, but these are minor details.

Still required on the edit view are the means to add top-
level datafields, as well as the ability to turn a single
data value into a new sub-level, a la JSON.

Also in this commit, ejs templates now carry an ejs file
extension, for syntax highlighting purposes.