Fix edge case for edit view
An edge case whereby unnecessary level corridors (vertical
bars) were being drawn for some data, in particular those
which are descended from the last child of some level
(including the top level) is fixed.  By adding an array
parameter to the recursive field creation function, this
seems to be rectified.  More testing is required to be sure,
obviously.