Refactor expandField to expand comma-list into multi-fields
Any datatier where the value (not the key) is a comma-
separated list will now expand into multiple fields, one
for each item in said list, each item becoming a new key
for a new datatier.  This is to reflect the future
behaviour of collapseLevel, which will take all keys in
the level and collapse them into the value field for the
parent level (into a comma-separated list, to be sure).