Implement collapseLevel function for JSON editor
The JSON editor can now collapse a level back into a key-
value pair, as per some (for now) odd but easy-to-implement
behaviour.  When collapsing, the function looks for all
children (and sub-children and so forth) of the calling
div and collapses their keys into a comma-separated list in
the new value field.  You lose all the values in the sub-
fields, but one can still expand the comma-separated list
in the value field back into a bunch of new keys for
sub-fields, albeit without any of the previous structural
information (i.e., it will all be in one sub-level).  More
complicated behaviour may be implemented in the future, but
don't hold your breath.