Fix deleteField case for last child deletion
The case whereby the field deleted is the last child of
its level and its previous sibling has children whose
corridors are broken by deletion has been addressed.
Now the deletion simply un-draws the corridors for the
level from which the field is being deleted, in the same
way these corridors are drawn when a new last child is
added after a (now penultimate) child who has its own
children.