Address edge cases for deleteField
Two edge cases have been addressed for deleteField one
whereby the field did not want to delete itself when its
parent has a certain number of children (a check done for
sublevels as to whether they only have one remaining child,
which is not necessary at the top level), and another
whereby the field did not want to delete itself as the only
remaining child of the datafields div.  This leads directly
into addressing the next edge case, which is adding a data
field to an empty datafields div.