Address empty document edge case for addField
The edge case whereby a new field was not being added to
a JSON document being edited when the document being edited
has no data fields in it (i.e., when the user deletes all
the data fields and wants to add a new field) has been
addressed.  A field is now created as expected when one
tries to add a field to an empty document.