Implement numeric adjustments for deleteField
The behaviour for deleteField has been updated such that
when a field is deleted, the function adjusts the numeric
identifiers for all siblings after the deleted field, and
for all those siblings' descendants as well.  This ensures
that all the numbers seen while editing the document are
dense, i.e., there are no numeric gaps between fields.

This behaviour seems correct so far but is as usual subject
to further testing.