Update README to reflect json-editor, name change
The README has been updated to reflect a host of changes
merged in by the json-editor feature, which sets the
project on a slightly new direction.  Project goals and
requirements have been somewhat refined but need more work.

Furthermore, the name of the project has been changed, and
a little blurb behind the name is now in the README.

Still needed in the README are a description of the project
directory structure, and a list of stuff that doesn't work
just yet.