Create end-to-end test for dashboard view
A single end-to-end test is now available for the dashboard
view, which only checks whether the app is working, more
specifically whether the app root returns a 200.  This test
is to be used to write other end-to-end view tests, hence
the (in this case) useless lines for JS coverage data.

It is worth noting at this early stage that testing with
puppeteer currently requires headful chrome as the app
rejects requests from headless chrome (as it well should).