Create tests for edit entry (JSON editor) view
This commit comprises the beginnings of the testing file
(more likely files plural) for the JSON editor view.  End-
to-end tests are minimal at this time, as the emphasis at
the moment is on display testing and testing the editor's
functionality (and adapting its behaviour).

Basic display testing right now creates a two level
complete quadtree and goes from there.  More complicated
starting points will be added later.  To actually get a
good long look at a display test, a tester will need to
insert a page.waitFor() statement at the end of said test
(just before the return statement).

Included in the commit are an updated Gruntfile which now
includes a watch task for firing off the aforementioned
tests, and the folding of the new entry view into the edit
entry view (any non-integer id prompts the user with a new
document now).