Configure coverage report generation using Grunt
Coverage data for js and css generated by Puppeteer is now
reported on, on a per-test basis, using
puppeteer-to-istanbul to generate report data, and nyc
(run by Grunt) to summarize said data.  Both a summary and
an in-depth html report are generated when a test file
is run.

Currently, when Grunt runs the task which fires off all
test files in succession, no coverage reports are generated
since coverage data is overwritten each time a test file
is run.  This behaviour is subject to change.

Owing to the increasing volume of terminal output when test
files are run, some Grunt terminal output, as well as some
choices for reporters for some test tasks in Gruntfile.js
have been adjusted.  Nyan.