Add links to dashboard view
Entries found on the dashboard are now linked directly
to the edit entry view, such that these dashboard entries
can be edited without having to find them using the
calendar.

In order to not mess up coverage reports for the edit entry
view, the tests for these dashboard links currently reside
in the edit entry view test file.