Add end-to-end test for accessing a doc from the calendar
An end-to-end test for accessing an extant JSON document
starting from the calendar is now available, and for now
resides in the edit entry view test file to preserve
coverage report veracity.  Whether end-to-end tests will
be moved to their own file (not concerned with coverage)
remains to be seen.