Extend data model such that data has ownership
Previously the data model was super simple and the only structure
imposed upon it was that it should be uniquely identified and
timestamped JSON.  Now there is a further requirement that data shall
each be associated with an (extant) user id.  This necessitated adding
to the database schema tables for users, roles, and (not strictly
required at this point, but for good measure) passwords.

To facilitate testing, a database destroying sql file has been included
so that one can quickly nuke the dummy data and refresh it with psql.