Add modal month select to month view
A means to change the month to a (somewhat) arbitrary choice of month
and year has been added.  Right now, only 2021-2023 are valid choices
for year, though the API still accepts all values between 1 and 2023.

Additionally, calendar has been renamed to month, since it may not
always be the case that the month is the only calendar view available.
The API still expects "calendar" as the resource name, in keeping with
the idea that there may be more calendar views to come.